This script includes growth defined by concentrations of substrates. I have modified Steve's original Thornley model (Thornley and Parsons 2014. J. Theor. Biol. 342: 1-14). In this orgininal version, growth is divided between plant compartments based on an allocation algorithm. This allocation algorithm has the drawback that a reduction in plant compartment size reduces allocatoin and growth of that compartment.
A reduction of growth of a compartment is fine however the model does not account for important source sink dynamics. That is, any organ can be and is both a source and sink of different substrates at different rates at different times.

I am attempting to resolve the difference between this original Thornley model's response to removal and compartment structural mass and substrates and observed allocation repsonsed.

To do this I used another of Thornley's models (Thornley 1998 modelling shoot-root relations. The only way forward (https://doi.org/10.1006/anbo.1997.0529)). Here growth is defined separately for each plant compartment, as mediated by substrate concentrations. In this model the transport of carbon substrate from leaves to roots and the transport of nitrogen from roots to leaves is explicitly conisdered. I extend this transport of substrates by making it bi-directional, i.e. if there is a substrate concentrtation gradient between any two compartments for any substrate then these substrates are transported at a rate determind by the difference in concentations (potential difference) and the resistance between the two target compartments. The resistance of each compartment is a function of their masses.

I'm currently investigating whether this method can work for our application.